pyproject.toml
[project]
name = "ipbes-classifier"
version = "0.1.0"
description = "IPBES multi-label text classifier for biodiversity research"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.5.2",
    "arxiv>=2.1.3",
    "autorank>=1.3.0",
    "beautifulsoup4>=4.13.3",
    "biopython>=1.85",
    # "bitsandbytes>=0.45.4",  # Temporarily commented out for macOS compatibility
    "clarivate-wos-starter-python-client",
    "datasets>=3.4.0",
    "deep-translator>=1.11.4",
    "deepspeed>=0.17.1",
    "dependencies>=7.7.1",
    "evaluate>=0.4.3",
    "faiss-cpu>=1.10.0",
    "gputil>=1.4.0",
    "gradio>=5.38.2",
    "hyperopt==0.2.5",
    "imblearn>=0.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "iterative-stratification>=0.1.9",
    "langchain-community>=0.3.20",
    "langchain-core>=0.3.49",
    "lxml>=5.3.1",
    "matplotlib>=3.10.1",
    "mistralai>=1.6.0",
    "mlflow>=2.22.0",
    "mlxtend>=0.23.4",
    "nlpaug>=1.1.11",
    "nltk>=3.9.1",
    "omegaconf>=2.3.0",
    # "onnxruntime>=1.21.0",  # Temporarily commented out for macOS compatibility
    "openai>=1.68.2",
    "orange3>=3.38.1",
    "pandas>=2.2.3",
    "protobuf>=6.30.2",
    "pyalex>=0.18",
    "pyyaml>=6.0.2",
    "ray[default,tune]>=2.44.1",
    "regex>=2024.11.6",
    "runtime>=0.1.4",
    "scikit-learn>=1.6.1",
    "scikit-posthocs>=0.11.4",
    "seaborn>=0.13.2",
    "sentence-transformers>=3.4.1",
    "sentencepiece>=0.2.0",
    "setuptools>=76.0.0",
    "spacy>=3.8.7",
    "streamlit>=1.47.0",
    "tenacity>=9.0.0",
    "tensorboard>=2.19.0",
    "tensorboardx>=2.6.2.2",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "transformers>=4.49.0",
    "virtualenv>=20.29.3",
    "wordcloud>=1.9.4",
]

[tool.uv.sources]
clarivate-wos-starter-python-client = { git = "https://github.com/clarivate/wosstarter_python_client.git" }
